requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.3.1
pandas==1.4.2
numpy==1.22.3
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
plotly==5.7.0
matplotlib==3.5.1
scikit-learn==1.0.2